+++ src/common/components/ConfirmationModal.less
@@ -1,4 +1,4 @@
-@import "../variables";
+@import "../../variables";
 
 .ConfirmationModal {
     padding: @modal-padding;
