+++ src/common/components/StickyAlertStripe.js
