+++ src/common/components/StickyAlertStripe.less
@@ -1,4 +1,4 @@
-@import '../variables';
+@import '../../variables';
 
 .StickyAlertStripe {
     position: fixed;
