+++ src/common/utils/capitalizeLocation.js
