+++ src/common/utils/getWorkLocation.js
