+++ src/favourites/Favourites.js
@@ -3,10 +3,10 @@ import { Column, Container, Row } from 'nav-frontend-grid';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
-import PageHeader from '../common/pageHeader/PageHeader';
-import DelayedSpinner from '../common/DelayedSpinner';
+import PageHeader from '../common/components/PageHeader';
+import DelayedSpinner from '../common/components/DelayedSpinner';
 import NotAuthenticated from '../authentication/NotAuthenticated';
-import RestoreScroll from '../common/RestoreScroll';
+import RestoreScroll from '../common/components/RestoreScroll';
 import NoUser from '../user/NoUser';
 import FavouriteAlertStripe from './alertstripe/FavouriteAlertStripe';
 import './Favourites.less';
