+++ src/favourites/list/FavouriteListItem.js
@@ -4,7 +4,7 @@ import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
 import { Lenkeknapp } from 'pam-frontend-knapper';
-import '../../common/Icons.less';
+import '../../common/components/Icons.less';
 import SearchResultsItemDetails from '../../search/searchResults/SearchResultsItemDetails';
 import { SHOW_MODAL_REMOVE_FROM_FAVOURITES } from '../favouritesReducer';
