+++ src/savedSearches/SavedSearches.js
@@ -3,8 +3,8 @@ import { Column, Container, Row } from 'nav-frontend-grid';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { connect } from 'react-redux';
-import PageHeader from '../common/pageHeader/PageHeader';
-import DelayedSpinner from '../common/DelayedSpinner';
+import PageHeader from '../common/components/PageHeader';
+import DelayedSpinner from '../common/components/DelayedSpinner';
 import NotAuthenticated from '../authentication/NotAuthenticated';
 import NoUser from '../user/NoUser';
 import SavedSearchAlertStripe from './alertstripe/SavedSearchAlertStripe';
