+++ src/savedSearches/list/SavedSearchListItem.js
@@ -6,7 +6,7 @@ import { connect } from 'react-redux';
 import { Link } from 'react-router-dom';
 import classNames from 'classnames';
 import { Flatknapp, Lenkeknapp } from 'pam-frontend-knapper';
-import '../../common/Icons.less';
+import '../../common/components/Icons.less';
 import { CONTEXT_PATH } from '../../fasitProperties';
 import { formatISOString, isValidISOString } from '../../utils';
 import NotifyTypeEnum from '../enums/NotifyTypeEnum';
