+++ src/search/Search.js
@@ -19,8 +19,8 @@ import Extent from './facets/Extent';
 import Occupations from './facets/Occupations';
 import Published from './facets/Published';
 import Sector from './facets/Sector';
-import DelayedSpinner from '../common/DelayedSpinner';
-import RestoreScroll from '../common/RestoreScroll';
+import DelayedSpinner from '../common/components/DelayedSpinner';
+import RestoreScroll from '../common/components/RestoreScroll';
 import './Search.less';
 import SearchBox from './searchBox/SearchBox';
 import { INITIAL_SEARCH, RESET_SEARCH, SEARCH } from './searchReducer';
