+++ src/search/searchResults/SearchResultsItemDetails.js
@@ -3,8 +3,8 @@ import { Normaltekst, Undertekst, Undertittel } from 'nav-frontend-typografi';
 import PropTypes from 'prop-types';
 import React from 'react';
 import { Link } from 'react-router-dom';
-import getEmployer from '../../common/getEmployer';
-import getWorkLocation from '../../common/getWorkLocation';
+import getEmployer from '../../common/utils/getEmployer';
+import getWorkLocation from '../../common/utils/getWorkLocation';
 import { CONTEXT_PATH } from '../../fasitProperties';
 import { formatISOString, isValidISOString } from '../../utils';
 import './SearchResultsItemDetails.less';
